wip, raw slice implementation
