+++ packages/fuels-core/src/function_selector.rs
@@ -71,7 +71,7 @@ fn resolve_arg(arg: &ParamType) -> String {
             let inner = resolve_arg(el_type);
             format!("s<{inner}>(s<{inner}>(rawptr,u64),u64)")
         }
-        ParamType::RawSlice => unimplemented!("`RawSlice` type is not supported as arguments"),
+        ParamType::RawSlice => "rawslice".to_string(),
         ParamType::Bytes => "s(s(rawptr,u64),u64)".to_string(),
     }
 }
